Add flags

+++ app/services/webhook_service.py
@@ -177,6 +177,10 @@ class WebhookService:
             )
             return None
 
+        if not runner_name.startswith('gcp-runner-'):
+            logger.warning("gcp-runner prefix not found in runner name %s. Ignoring job.", runner_name)
+            return
+
         try:
             self.gcloud_client.delete_runner_instance(
                 runner_name, delivery_id=delivery_id
